app/requirements.txt
Flask==2.0.2
google-cloud-firestore==2.3.4
